+++ medusa/metadata/wdtv.py
@@ -67,16 +67,16 @@ class WDTVMetadata(generic.GenericMetadata):
         self.poster_name = 'folder.jpg'
 
         # web-ui metadata template
-        self.eg_show_metadata = '<i>not supported</i>'
+        # self.eg_show_metadata = '<i>not supported</i>'
         self.eg_episode_metadata = 'Season##\\<i>filename</i>.xml'
-        self.eg_fanart = '<i>not supported</i>'
+        # self.eg_fanart = '<i>not supported</i>'
         self.eg_poster = 'folder.jpg'
-        self.eg_banner = '<i>not supported</i>'
+        # self.eg_banner = '<i>not supported</i>'
         self.eg_episode_thumbnails = 'Season##\\<i>filename</i>.metathumb'
         self.eg_season_posters = 'Season##\\folder.jpg'
-        self.eg_season_banners = '<i>not supported</i>'
-        self.eg_season_all_poster = '<i>not supported</i>'
-        self.eg_season_all_banner = '<i>not supported</i>'
+        # self.eg_season_banners = '<i>not supported</i>'
+        # self.eg_season_all_poster = '<i>not supported</i>'
+        # self.eg_season_all_banner = '<i>not supported</i>'
 
     # Override with empty methods for unsupported features
     def retrieveShowMetadata(self, folder):