+++ themes-default/slim/src/store/modules/index.js
@@ -1,0 +1,9 @@
+export { default as auth } from './auth';
+export { default as config } from './config';
+export { default as defaults } from './defaults';
+export { default as metadata } from './metadata';
+export { default as notifications } from './notifications';
+export { default as qualities } from './qualities';
+export { default as shows } from './shows';
+export { default as socket } from './socket';
+export { default as statuses } from './statuses';