+++ themes-default/slim/static/js/config/notifications.js
@@ -472,7 +472,7 @@ MEDUSA.config.notifications = function() { // eslint-disable-line max-lines
         host = host.length > 0 ? host : null;
         let port = $('#email_port').val();
         port = port.length > 0 ? port : null;
-        const tls = $('#email_tls').attr('checked') === undefined ? 0 : 1;
+        const tls = $('#email_tls').is(':checked') ? 1 : 0;
         let from = $('#email_from').val();
         from = from.length > 0 ? from : 'root@localhost';
         const user = $('#email_user').val().trim();